make another component for agent modal content to keep it away from client component
