+++ web-app/src/components/agents/agent-modal/index.ts
@@ -1,1 +1,2 @@
 export * from "./agent-modal";
+export * from "./agent-modal-content";
